pyproject.toml
[tool.poetry]
name = "rag-from-scratch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-community = "^0.3.0"
tiktoken = "^0.7.0"
langchain-openai = "^0.2.0"
langchainhub = "^0.1.21"
chromadb = "^0.5.5"
langchain = "^0.3.0"
jupyter = "^1.1.1"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
langfuse = "^2.48.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
